requirements.txt
line-bot-sdk
flask
configparser
azure-ai-language-conversations
